+++ paperpi/library/run_module.py
@@ -7,6 +7,7 @@
 import importlib
 import inspect
 from pathlib import Path
+from datetime import datetime
 
 
 
@@ -18,7 +19,7 @@ def run_module(module_args=[]):
         print('Usage:')
         print('--run_plugin_func plugin.function [ARG1] [ARG2]')
         print('\nor for more information:\n--plugin_info plugin')
-        
+         
     if not module_args:
         print_usage()
         return
@@ -50,6 +51,62 @@ def run_module(module_args=[]):
     except Exception as e:
         print(f'error: {e}')
         return
+    
+
+
+
+
+
+
+def add_config(module=None, config_file=None):
+    def print_usage():
+        print('Adds basic configuration to config file')
+        print('Usage:')
+        print('--add_config plugin user|daemon')
+        print('\nExample: --add_config moon_phase daemon')
+        print('\nfor list of plugins:\n--list_plugins')
+    
+    def time_stamp():
+        return datetime.now().strftime('# {} added >>>>>>  %Y.%m.%d %H:%M:%S\n')    
+    
+    if not module:
+        print_usage()
+        return
+    
+    my_module = module
+    my_config = Path(config_file)
+    
+    try:
+        i = importlib.import_module(f'plugins.{module}.constants')
+    except Exception as e:
+        print(f'plugin "{module}" not found\n')
+        print_usage()
+        return
+    
+    try:
+        config = i.sample_config
+    except AttributeError:
+        print('this plugin does not appear to have a sample configuration')
+        print('aborting')
+        return
+
+    try:
+        with open(config_file, 'a') as f:
+            f.write(time_stamp().format('start'))
+            f.write(config)
+            f.write(time_stamp().format('end'))
+    except PermissionError:
+        print('It appears you are trying to append to the daemon config file;')
+        print('you may need to run this with "sudo."')
+        return
+    except OSError as e:
+        print(f'An error occured while writing file "{config}": {e}')
+        return
+
+    print(f'Finished writing configuration for {module}')
+    print(f'It is a very good idea to open {config_file}\nand check the configuration you have just added!')
+    
+